requirements.txt
asgiref==3.6.0
Django==4.1.6
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
